Move VariableFractionTransformer to custom module
