+++ oemof/solph/__init__.py
@@ -3,9 +3,7 @@
 
 """
 from oemof.solph.network import (Sink, Source, LinearTransformer, Bus,
-                                 Flow, EnergySystem, LinearN1Transformer,
-                                 VariableFractionTransformer)
-
+                                 Flow, EnergySystem, LinearN1Transformer)
 from oemof.solph.models import OperationalModel
 from oemof.solph.groupings import GROUPINGS
 from oemof.solph.options import (Investment, BinaryFlow, DiscreteFlow)
