+++ oemof/solph/groupings.py
@@ -17,8 +17,7 @@ groupings specified like this:
 
 """
 
-from .network import (Bus, LinearTransformer, LinearN1Transformer,
-                      VariableFractionTransformer)
+from .network import (Bus, LinearTransformer, LinearN1Transformer)
 from .options import Investment
 from . import blocks
 import oemof.groupings as groupings
@@ -42,8 +41,6 @@ def constraint_grouping(node):
     # constraints are grouped by overriding the method in future subclasses.
     if isinstance(node, Bus) and node.balanced:
         return blocks.Bus
-    if isinstance(node, VariableFractionTransformer):
-        return blocks.VariableFractionTransformer
     if isinstance(node, LinearTransformer):
         return blocks.LinearTransformer
     if isinstance(node, LinearN1Transformer):
